replace [NodeQuery] with NodePath
previously we were using NodeQuery but we are going to need to access more
things at specific points in the Node tree. To make this code easier to work
with I decided to replace NodeQuery with NodeSelector and [NodeQuery] with
NodePath.

NodePath is a reversed list of selectors, meaning that consing a selector to a
NodePath would add it to the end of the Path and foldr over that would begin at
the outermost selector.