WIP: pass NodeCursor while formatting the Node tree
As I walk through the Node tree to run formatNode on each node I prepend a
breadcrumb the to the cursor. This will be used to check if any rules configured
in the jbeam_formmatting_rules.jbfl is applicable for the point the cursor is pointing at.

jbfl is the preliminary file name extension for the new mini language that I am
planning to write to represent formatting rules.

A formatting rule could look like this

.*.nodes:filter(isVerticeNode) {
  justify-amount : 5;
}

.*.nodes:filter(isVerticeNode) works as follows:

    .* is a wildcard that matches any key inside an object.

    .nodes selects the "nodes" section inside the matched object.

    filter(isVerticeNode) filters the nodes, applying rules in below the coming
    block for only those nodes that are identified as vertices.

Inside the block, justify-amount : 5; sets a rule: scalars (like numbers or strings) will be padded with spaces to be at least 5 characters long.