Always read shipped config file (#133)
Upon installation a configuration file will always be created if missing. But if a user already installed jbeam-edit, a configuration was created, then the user updates to a new version, then the new installation would use older configuration which was in AppData. This patch will load the shipped configuration file and the AppData version if it exists, preferring the the AppData settings if availabale.

Since jbeam-edit now both loads a config file in AppData as well as the shipped minimal configuration, I added a test which proves that combining the configurations at runtime works. I also improved logging in the test suite to be able debug issues easier.